backend/requirements.txt
fastapi
uvicorn
textblob
pyttsx3
pyyaml